TINKERPOP-2079 Moved RemoteGraph to gremlin-test
This is a breaking change as it removes the long deprecated RemoteGraph from gremlin-core. Users should not have been using this at this point, so the impact should be fairly limited. RemoteGraph is still needed as a test component because we need a Graph instance for the JVM test suite to run properly. This change required some refactoring as well in the RemoteGraphProvider so that configurations were being passed around properly to allow for the test suite to re-attach elements. That's yet another problem with the JVM test suite - it unfotunately assumes embedded Graph assertions.